arcreg/requirements.txt
dj-database-url==0.4.2
Django==1.11.3
django-crispy-forms==1.6.1
django-material==1.0.0
git+git://github.com/CodeFanatic23/django-registration@login
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
openpyxl==2.4.8
psycopg2==2.7.1
pytz==2017.2
whitenoise==3.3.0
xlrd==1.0.0
xlwt==1.2.0
